SpeechToTextAnywhere/requirements.txt
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
pyautogui>=0.9.54
keyboard>=0.13.5
pystray>=0.19.4
Pillow>=10.0.0
PyInstaller>=6.0.0